--- Dionysos.Database/Class1.cs
@@ -1,5 +1,0 @@
-﻿namespace Dionysos.Database;
-
-public class Class1
-{
-}
