+++ Dionysos/Dionysos.csproj
@@ -31,7 +31,6 @@
 
     <ItemGroup>
         <Folder Include="Dionysos.App" />
-        <Folder Include="Dionysos.Framework" />
     </ItemGroup>
 
     <ItemGroup>
@@ -41,6 +40,7 @@
     </ItemGroup>
 
     <ItemGroup>
+      <ProjectReference Include="..\Dionysos.BL\Dionysos.BL.csproj" />
       <ProjectReference Include="..\Dionysos.Database\Dionysos.Database.csproj" />
     </ItemGroup>
