+++ Dionysos/Extensions/InventoryItemDtoExtensions.cs
@@ -4,7 +4,7 @@ using Google.Protobuf.WellKnownTypes;
 
 namespace Dionysos.Dionysos.BL.Extensions;
 
-public static class InventoryItemExtensions
+public static class InventoryItemDtoExtensions
 {
     public static InventoryItem ToProtobufItem(this InventoryItemDto itemDto)
     {
@@ -13,15 +13,6 @@ public static class InventoryItemExtensions
             Id = itemDto.Id,
             BestBefore = Timestamp.FromDateTime(itemDto.BestBefore ?? DateTime.MinValue),
             Ean = itemDto.Ean
-        };
-    }
-
-    public static Dionysos.Database.Database.InventoryItem ToDbInventoryItem(this InventoryItemDto inventoryItemDto)
-    {
-        return new Dionysos.Database.Database.InventoryItem
-        {
-            BestBefore = inventoryItemDto.BestBefore,
-            Ean = inventoryItemDto.Ean
         };
     }
 }
