+++ Dionysos/Extensions/ProtobufArticleExtensions.cs
