+++ Dionysos/Extensions/ProtobufInventoryItemExtensions.cs
